+++ response.go
@@ -59,6 +59,9 @@ func (r *Response) FileName() string {
 }
 
 func (r *Response) fixCharset(detectCharset bool, defaultEncoding string) error {
+	if len(r.Body) == 0 {
+		return nil
+	}
 	if defaultEncoding != "" {
 		tmpBody, err := encodeBytes(r.Body, "text/plain; charset="+defaultEncoding)
 		if err != nil {
@@ -68,6 +71,16 @@ func (r *Response) fixCharset(detectCharset bool, defaultEncoding string) error 
 		return nil
 	}
 	contentType := strings.ToLower(r.Headers.Get("Content-Type"))
+
+	if strings.Contains(contentType, "image/") ||
+		strings.Contains(contentType, "video/") ||
+		strings.Contains(contentType, "audio/") ||
+		strings.Contains(contentType, "font/") {
+		// These MIME types should not have textual data.
+
+		return nil
+	}
+
 	if !strings.Contains(contentType, "charset") {
 		if !detectCharset {
 			return nil